requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pyyaml>=6.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
cachetools>=5.3.0
structlog>=23.2.0
prometheus-client>=0.19.0
redis>=5.0.0
aiofiles>=23.2.0
